while deleting return the deleted item from the febric
